rbd: add rbd object-map check
Add a CLI option to verify the object map for an unmapped image
for debugging purposes. Syntax:

rbd object-map check <image-name>

This operation is not supported on currently mapped images.

Any inconsistencies that could affect the correctness of future
operations are noted and the object map (and fast diff state, if
enabled) is invalidated. Inconsistencies that do not affect
correctness will be emitted at debug level 1 or higher.

Consolidate code path with object-map rebuild and remove newly
unused code.

Fixes: http://tracker.ceph.com/issues/14867
Signed-off-by: Douglas Fuller <dfuller@redhat.com>